Added tableHeader element.
